refactor: mv ChannelList
